Improved Canvas: Using float params for transformation affected drawing methods.
Core API changes:
  * replaced 'int' with 'float' in all ch.jeda.ui.Canvas.draw...() methods
  * replaced 'int' with 'float' in all ch.jeda.ui.Canvas.fill...() methods
  + added method ch.jeda.ui.Canvas.drawPolygon(float[])
  + added method ch.jeda.ui.Canvas.fillPolygon(float[])
  - removed public field ch.jeda.ui.Color.value
  + added method ch.jeda.ui.Color.getValue()

Platform API changes (ch.jeda.platform):
  * replaced 'int' with 'float' in all CanvasImp.draw...() methods
  * replaced 'int' with 'float' in all CanvasImp.fill...() methods

+++ core/src/ch/jeda/platform/CanvasImp.java
@@ -23,25 +23,25 @@ public interface CanvasImp {
 
     void copyFrom(int x, int y, CanvasImp source);
 
-    void drawCircle(int x, int y, int radius);
+    void drawCircle(float x, float y, float radius);
 
-    void drawImage(int x, int y, ImageImp image, int alpha);
+    void drawImage(float x, float y, ImageImp image, int alpha);
 
-    void drawLine(int x1, int y1, int x2, int y2);
+    void drawLine(float x1, float y1, float x2, float y2);
 
-    void drawPolygon(int[] points);
+    void drawPolygon(float[] points);
 
-    void drawRectangle(int x, int y, int width, int height);
+    void drawRectangle(float x, float y, float width, float height);
 
-    void drawText(int x, int y, String text);
+    void drawText(float x, float y, String text);
 
     void fill();
 
-    void fillCircle(int x, int y, int radius);
+    void fillCircle(float x, float y, float radius);
 
-    void fillPolygon(int[] points);
+    void fillPolygon(float[] points);
 
-    void fillRectangle(int x, int y, int width, int height);
+    void fillRectangle(float x, float y, float width, float height);
 
     int getHeight();
